Make placeholder dependency use hedvig DS
Split into a core and a Hedvig module to split the core functionality
and the styling, as the original implementation worked too.
