+++ app/ui/placeholder/placeholder-core/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+  id("hedvig.gradle.plugin")
+  id("hedvig.android.library")
+}
+
+hedvig {
+  compose()
+}
+
+dependencies {
+  implementation(libs.androidx.compose.animationCore)
+  implementation(libs.androidx.compose.runtime)
+  implementation(libs.androidx.compose.uiUtil)
+}
