+++ app/ui/placeholder/placeholder-core/src/main/AndroidManifest.xml
