+++ app/ui/placeholder/placeholder-hedvig/build.gradle.kts
@@ -8,8 +8,9 @@ hedvig {
 }
 
 dependencies {
+  api(projects.placeholderCore)
   implementation(libs.androidx.compose.animationCore)
-  implementation(libs.androidx.compose.material3)
   implementation(libs.androidx.compose.runtime)
   implementation(libs.androidx.compose.uiUtil)
+  implementation(projects.designSystemHedvig)
 }
