+++ app/ui/ui-emergency/build.gradle.kts
@@ -18,5 +18,5 @@ dependencies {
   implementation(projects.coreUiData)
   implementation(projects.dataContractAndroid)
   implementation(projects.designSystemHedvig)
-  implementation(projects.placeholder)
+  implementation(projects.placeholderHedvig)
 }
